[AIP-21] Move operators.gcs_to_s3 providers.amazon.aws.operators.gcs_to_s3
