+++ tests/providers/amazon/aws/operators/test_gcs_to_s3.py
@@ -21,8 +21,8 @@ import unittest
 
 import mock
 
-from airflow.operators.gcs_to_s3 import GCSToS3Operator
 from airflow.providers.amazon.aws.hooks.s3 import S3Hook
+from airflow.providers.amazon.aws.operators.gcs_to_s3 import GCSToS3Operator
 
 try:
     from moto import mock_s3
@@ -42,7 +42,7 @@ class TestGCSToS3Operator(unittest.TestCase):
     # Test1: incremental behaviour (just some files missing)
     @mock_s3
     @mock.patch('airflow.gcp.operators.gcs.GCSHook')
-    @mock.patch('airflow.operators.gcs_to_s3.GCSHook')
+    @mock.patch('airflow.providers.amazon.aws.operators.gcs_to_s3.GCSHook')
     def test_execute_incremental(self, mock_hook, mock_hook2):
         mock_hook.return_value.list.return_value = MOCK_FILES
         mock_hook.return_value.download.return_value = b"testing"
@@ -72,7 +72,7 @@ class TestGCSToS3Operator(unittest.TestCase):
     # Test2: All the files are already in origin and destination without replace
     @mock_s3
     @mock.patch('airflow.gcp.operators.gcs.GCSHook')
-    @mock.patch('airflow.operators.gcs_to_s3.GCSHook')
+    @mock.patch('airflow.providers.amazon.aws.operators.gcs_to_s3.GCSHook')
     def test_execute_without_replace(self, mock_hook, mock_hook2):
         mock_hook.return_value.list.return_value = MOCK_FILES
         mock_hook.return_value.download.return_value = b"testing"
@@ -103,7 +103,7 @@ class TestGCSToS3Operator(unittest.TestCase):
     # Test3: There are no files in destination bucket
     @mock_s3
     @mock.patch('airflow.gcp.operators.gcs.GCSHook')
-    @mock.patch('airflow.operators.gcs_to_s3.GCSHook')
+    @mock.patch('airflow.providers.amazon.aws.operators.gcs_to_s3.GCSHook')
     def test_execute(self, mock_hook, mock_hook2):
         mock_hook.return_value.list.return_value = MOCK_FILES
         mock_hook.return_value.download.return_value = b"testing"
@@ -132,7 +132,7 @@ class TestGCSToS3Operator(unittest.TestCase):
     # Test4: Destination and Origin are in sync but replace all files in destination
     @mock_s3
     @mock.patch('airflow.gcp.operators.gcs.GCSHook')
-    @mock.patch('airflow.operators.gcs_to_s3.GCSHook')
+    @mock.patch('airflow.providers.amazon.aws.operators.gcs_to_s3.GCSHook')
     def test_execute_with_replace(self, mock_hook, mock_hook2):
         mock_hook.return_value.list.return_value = MOCK_FILES
         mock_hook.return_value.download.return_value = b"testing"
@@ -163,7 +163,7 @@ class TestGCSToS3Operator(unittest.TestCase):
     # Test5: Incremental sync with replace
     @mock_s3
     @mock.patch('airflow.gcp.operators.gcs.GCSHook')
-    @mock.patch('airflow.operators.gcs_to_s3.GCSHook')
+    @mock.patch('airflow.providers.amazon.aws.operators.gcs_to_s3.GCSHook')
     def test_execute_incremental_with_replace(self, mock_hook, mock_hook2):
         mock_hook.return_value.list.return_value = MOCK_FILES
         mock_hook.return_value.download.return_value = b"testing"
